Create stage 3 code
Stage 3 of the development of code for my interactive tic tac toe project. This part entails checking what is currently on the board to assess where in the game we're at i.e. a win, loss, draw, or still in play, and whether a move on the board is currently possible or impossible.